+++ src/main/kotlin/app/revanced/patches/youtube/layout/sponsorblock/resource/patch/SponsorBlockResourcePatch.kt
@@ -23,9 +23,6 @@ import app.revanced.util.resources.ResourceUtils.mergeStrings
 @DependsOn([FixLocaleConfigErrorPatch::class, SettingsPatch::class, ResourceMappingPatch::class])
 @Version("0.0.1")
 class SponsorBlockResourcePatch : ResourcePatch {
-    companion object {
-        internal var reelButtonGroupResourceId: Long = 0
-    }
 
     override fun execute(context: ResourceContext): PatchResult {
         val youtubePackage = "com.google.android.youtube"
@@ -106,10 +103,6 @@ class SponsorBlockResourcePatch : ResourcePatch {
                 break
             }
         }.close() // close afterwards
-
-        reelButtonGroupResourceId = ResourceMappingPatch.resourceMappings.single {
-            it.type == "id" && it.name == "reel_persistent_edu_button_group"
-        }.id
 
         return PatchResultSuccess()
     }